currencies dialog: a currency left without a name is not saved away
While its name is retyped, one of the tour's currencies keeps its rate and its ✕, and
saving it nameless is refused - removing is what ✕ is for; saved, it used to go, its
expenses converted, on the strength of an empty box. A reference that was a currency just
added and got emptied is replaced. Clippy's leftovers from the i64 change.

+++ rust/crates/tc-core/src/units.rs
@@ -249,7 +249,7 @@ pub fn cents_sibling<'a>(tour: &'a Tour, of: &CurrencyId) -> Option<&'a Currency
 /// goes by its name instead: the ones people count in cents.
 pub fn cents_by_default(rate: i64, name: &str, others: &[i64]) -> bool {
     match others.iter().copied().filter(|r| *r > 0).min() {
-        Some(cheapest) if cheapest < rate => rate as i64 >= 30 * cheapest as i64,
+        Some(cheapest) if cheapest < rate => rate >= 30 * cheapest,
         _ => {
             let n = name.trim().to_uppercase();
             ["EUR", "USD", "BAM", "BGN", "GBP", "CHF", "ЕВРО", "EURO", "KM", "LEV"]
